Log slide manifest when training MIL models
